refactor(manager): clean up file struct, add todo comments
