+++ lua/huez-manager/api/init.lua
@@ -1,0 +1,5 @@
+return {
+  colorscheme = require("huez-manager.api.colorscheme"),
+  live = require("huez-manager.api.live"),
+  favorites = require("huez-manager.api.favorites"),
+}
