+++ lua/huez-manager/api/live.lua
@@ -1,4 +1,4 @@
-local api_utils = require("huez_manager.utils")
+local api_utils = require("huez-manager.utils.api_utils")
 
 local M = {}
 
@@ -12,7 +12,7 @@ M.installed = api_utils.load_themes("live")
 M.add = function(theme)
   -- vim.notify("Adding " .. theme .. " to live themes")
   api_utils.add_theme("live", theme)
-  M.live_themes = api_utils.load_themes("live")
+  M.installed = api_utils.load_themes("live")
 end
 
 --- Removes a theme from the `huez-live-themes` file.
@@ -21,7 +21,7 @@ end
 ---@return nil
 M.remove = function(theme)
   api_utils.remove_theme("live", theme)
-  M.live_themes = api_utils.load_themes("live")
+  M.installed = api_utils.load_themes("live")
 end
 
 return M
