+++ lua/huez/pickers/themes/actions.lua
@@ -1,7 +1,7 @@
 local telescope_actions = require("telescope.actions")
 local telescope_actions_state = require("telescope.actions.state")
-local api = require("huez_manager.api")
-local log = require("huez.utils.log")
+local api = require("huez-manager.api.init")
+local log = require("huez-manager.utils.log")
 
 local M = {}
