--- lua/huez_manager/api/init.lua
@@ -1,5 +1,0 @@
-return {
-  colorscheme = require("huez_manager.api.colorscheme"),
-  live = require("huez_manager.api.live"),
-  favorites = require("huez_manager.api.favorites"),
-}
